task-01 ajustes e organização do projeto
