+++ Shared/Models/Countdown.swift
@@ -1,0 +1,14 @@
+//
+//  GoalTime.swift
+//  Countdown (iOS)
+//
+//  Created by Leonir Alves Deolindo on 24/08/20.
+//
+
+import Foundation
+
+struct Countdown {
+    let hours: Int
+    let minutes: Int
+    let seconds: Int
+}
